improve modularization of src/github_bot
